fix: user edit
Editing a user without changing their password would fail with the error "Property [password] does not exist on this collection instance". This change uses the `$record` for the user as provided by filament instead of fetching the user based on form input data, since the user will never be found if you search for it in the DB by the new name which was not yet persisted.